+++ src/ffi/ffi_fastica.rs
@@ -5,8 +5,8 @@ pub extern "C" fn fast_ica(
     whitened_matrix: MatrixHandle,
     n_components: u64,
     max_iter: u64,
-    conv_threshold: f32,
-    alpha: f32,
+    conv_threshold: f64,
+    alpha: f64,
     cfid: u32,
 ) -> MatrixHandle {
